refactor: extract insights time selector
