+++ frontend/src/insights/InsightsPage.tsx
@@ -48,7 +48,10 @@ import {
   type CategoryDriver,
   type CategoryTrendSection,
 } from './components/IncomeExpenseBreakdownCard'
-import { InsightsRangeSelector, type InsightsRangeSelectorOption } from './components/InsightsRangeSelector'
+import {
+  InsightsFloatingRangeControl,
+  type InsightsRangePreset,
+} from './components/InsightsFloatingRangeControl'
 import {
   MerchantDistributionCard,
   type MerchantMarketTile,
@@ -73,8 +76,6 @@ import {
   SavingsRateTrendCard,
   type SavingsRateHistoryPoint,
 } from './components/SavingsRateTrendCard'
-
-type InsightsRangePreset = 'THIS_WEEK' | 'THIS_MONTH' | 'THIS_YEAR' | 'LAST_WEEK' | 'LAST_MONTH' | 'CUSTOM'
 
 type MerchantMarketEntry = {
   id: string
@@ -132,15 +133,6 @@ function useInsightCardVisibility() {
 
   return [ref, isVisible] as const
 }
-
-const INSIGHTS_RANGE_OPTIONS: InsightsRangeSelectorOption<InsightsRangePreset>[] = [
-  { value: 'THIS_WEEK', label: 'WTD', description: 'This week' },
-  { value: 'THIS_MONTH', label: 'MTD', description: 'This month' },
-  { value: 'THIS_YEAR', label: 'YTD', description: 'This year' },
-  { value: 'LAST_WEEK', label: 'LW', description: 'Last week' },
-  { value: 'LAST_MONTH', label: 'LM', description: 'Last month' },
-  { value: 'CUSTOM', label: 'Custom' },
-]
 
 const EMPTY_FLOW_ENTRIES: InsightsFlowEntry[] = []
 
@@ -670,106 +662,6 @@ function getMerchantMarketLayout(merchants: MerchantMarketEntry[]): MerchantMark
   )
 }
 
-function InsightsFloatingRangeControl({
-  preset,
-  customFrom,
-  customTo,
-  customInvalid,
-  onPresetChange,
-  onCustomFromChange,
-  onCustomToChange,
-}: {
-  preset: InsightsRangePreset
-  customFrom: string
-  customTo: string
-  customInvalid: boolean
-  onPresetChange: (value: InsightsRangePreset) => void
-  onCustomFromChange: (value: string) => void
-  onCustomToChange: (value: string) => void
-}) {
-  const inputDates = getRangeInputDates(preset, customFrom, customTo)
-
-  function handleCustomFromChange(value: string) {
-    if (preset !== 'CUSTOM') {
-      onCustomToChange(inputDates.to)
-      onPresetChange('CUSTOM')
-    }
-    onCustomFromChange(value)
-  }
-
-  function handleCustomToChange(value: string) {
-    if (preset !== 'CUSTOM') {
-      onCustomFromChange(inputDates.from)
-      onPresetChange('CUSTOM')
-    }
-    onCustomToChange(value)
-  }
-
-  const dateFields = (
-    <div>
-      <div className="grid grid-cols-[minmax(0,1fr)_auto_minmax(0,1fr)] items-center gap-2">
-        <input
-          type="date"
-          className={`app-input app-date-input-balanced min-w-0 ${customInvalid ? 'app-input-error' : ''}`}
-          aria-label="Insights start date"
-          value={inputDates.from}
-          onChange={(event) => handleCustomFromChange(event.target.value)}
-        />
-        <span className="text-xs font-semibold uppercase" style={{ color: 'var(--app-text-subtle)' }}>
-          to
-        </span>
-        <input
-          type="date"
-          className={`app-input app-date-input-balanced min-w-0 ${customInvalid ? 'app-input-error' : ''}`}
-          aria-label="Insights end date"
-          value={inputDates.to}
-          onChange={(event) => handleCustomToChange(event.target.value)}
-        />
-      </div>
-      {customInvalid && (
-        <p className="mt-1 text-xs" style={{ color: 'var(--app-negative)' }}>
-          Start date must be on or before end date.
-        </p>
-      )}
-    </div>
-  )
-
-  const renderControl = (dropdownPlacement?: 'bottom' | 'top') => (
-    <div className="app-card rounded-xl p-3">
-      <InsightsRangeSelector
-        value={preset}
-        options={INSIGHTS_RANGE_OPTIONS}
-        onChange={onPresetChange}
-        ariaLabel="Insights date range"
-        className="w-full"
-        sheetTitle="Insights date range"
-        dropdownPlacement={dropdownPlacement}
-      />
-      <div className="mt-2">
-        {dateFields}
-      </div>
-    </div>
-  )
-
-  return (
-    <>
-      <div className="pointer-events-none fixed inset-x-4 bottom-4 z-20 min-[760px]:hidden">
-        <div className="pointer-events-auto">
-          {renderControl('top')}
-        </div>
-      </div>
-
-      <div className="pointer-events-none absolute inset-x-0 bottom-0 top-0 z-40 hidden min-[760px]:block">
-        <div className="sticky top-6 flex justify-end">
-          <div className="pointer-events-auto w-[24rem]">
-            {renderControl()}
-          </div>
-        </div>
-      </div>
-    </>
-  )
-}
-
 function SectionHeader({
   icon: Icon,
   label,
@@ -920,7 +812,7 @@ export default function InsightsPage() {
   )
   return (
     <div className="relative">
-      <header className="app-page-header min-[760px]:pr-[25rem]">
+      <header className="app-page-header min-[1050px]:pr-[25rem]">
         <h1 className="app-page-title">Insights</h1>
         <p className="app-page-description">
           See where your money goes, spot patterns, and take control with confidence.
@@ -928,15 +820,15 @@ export default function InsightsPage() {
       </header>
       <InsightsFloatingRangeControl
         preset={rangePreset}
-        customFrom={customFrom}
-        customTo={customTo}
+        fromDateValue={rangeInputDates.from}
+        toDateValue={rangeInputDates.to}
         customInvalid={customInvalid}
         onPresetChange={setRangePreset}
         onCustomFromChange={setCustomFrom}
         onCustomToChange={setCustomTo}
       />
 
-      <div className="space-y-4">
+      <div className="space-y-4 pb-28 min-[1050px]:pb-0">
         <div ref={periodGlanceCardRef}>
           <PeriodGlanceCard
             header={<SectionHeader icon={Sparkles} label="This Period at a Glance" />}
